backend/requirements.txt
Flask==2.0.2
Flask-Mail==0.9.1
Flask-CORS==3.0.10
mysql-connector-python==8.0.28
Werkzeug==2.0.3
boto3==1.18.68
